Wayland: split seat handling into src/wayland/wl_seat.c
